Move templates pkg to machine:
The functionality didnt need its own package.
